Add menu state
Note the tricks required to get it to work properly, since the
canvas that is visible does not handle mouse events.
The trick is to hide the game canvas first, and enlarge the
original canvas. After clicking, hide the original canvas, and
show the game canvas.
Probably easier to just read the code in MenuState.js.